requirements.txt
# Copyright (C) 2019 Cognizant Digital Business, Cognizant Artifical Intelligence Practice.  All rights reserved.

gpflow==1.3.0
grpcio-tools==1.16.0
grpcio==1.16.0
protobuf==3.9.1
